accessible/tests/mochitest/treeupdate/test_bug884251.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<head>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

<script>

function boom() {
  document.getElementById("k").removeAttribute("href");
  ok(true, "changing iframe contents doesn't cause assertions");
  SimpleTest.finish();
}

</script>
</head>

<body onload="boom();">
<iframe src="data:text/html,1"><link id="k" href="data:text/html,2" /></iframe>
</body>
</html>
